requirements.txt
Django==5.0.8
djangorestframework==3.15.2
python-dotenv==1.0.1
Pillow==10.4.0
python-barcode==0.15.1
reportlab==4.2.2
